Change pattern used to start threads
With the previous pattern, every other WebConnectivity test
hanged because the count of running threads was not correctly
updated. With the current pattern, things seem to work in a
reliable way after running tests several times.

I am not sure I understand what changed, except that the new
pattern is equal to the one used by measurement-kit/libndt,
which recently got a security review where that pattern was
not flagged as suspicious.

I am leaving myself a note to double check what happened and
try to further understand what was going on. For now, and for
start testing this micro-library with MK, this is enough.